third_party/WebKit/LayoutTests/http/tests/credentialmanager/federatedcredential-basics.html

<!DOCTYPE html>
<title>Credential Manager: FederatedCredential basics.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="/serviceworker/resources/interfaces.js"></script>
<script>
test(function() {
    var credential = new FederatedCredential('id', 'name', 'https://example.com/avatar.png', 'https://federation.net/');

    assert_true(credential instanceof Credential);
    assert_true(credential instanceof FederatedCredential);

    verifyInterface('FederatedCredential', credential, {
        id: 'string',
        name: 'string',
        avatarURL: 'string',
        federation: 'string'
    });

    assert_equals(credential.id, 'id');
    assert_equals(credential.name, 'name');
    assert_equals(credential.avatarURL, 'https://example.com/avatar.png');
    assert_equals(credential.federation, 'https://federation.net/');
}, 'Interfaces and attributes of FederatedCredential');

test(function() {
    assert_throws(new SyntaxError(), function () {
        var credential = new FederatedCredential('id', 'name', '-', 'https://federation.net/');
    });
}, 'Construct a FederatedCredential with an invalid avatar URL.');

test(function() {
    assert_throws(new SyntaxError(), function () {
        var credential = new FederatedCredential('id', 'name', 'https://example.com/avatar.png', '-');
    });
}, 'Construct a FederatedCredential with an invalid federation URL.');
</script>
